Split watcher and host logic
This change allows to use multiple instances of the plugin and even
work with external watchers depending on how well they designed.

One of the running plugins will be a "master" plugin which runs a
watcher service while other plugins will use this service. One of
these plugins will become a "master" plugin in the moment when
previous "master" plugin was removed.